pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cirkelline"
version = "1.3.5"
description = "Multi-agent AI orchestration system - Personal assistant with specialist AI agents"
readme = "README.md"
license = {text = "Proprietary"}
requires-python = ">=3.11"
authors = [
    {name = "Cirkelline Team", email = "dev@cirkelline.com"}
]
keywords = ["ai", "agents", "orchestration", "fastapi", "gemini"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "agno>=2.3.4",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-dotenv",
    "google-genai>=1.52.0",
    "sqlalchemy",
    "pgvector",
    "psycopg[binary]",
    "pypdf",
    "python-docx",
    "beautifulsoup4",
    "markdown",
    "duckduckgo-search",
    "ddgs",
    "exa_py",
    "tavily-python",
    "psutil",
    "pyjwt>=2.8.0",
    "bcrypt>=4.1.0",
    "cryptography>=43.0.0",
    "tzlocal==5.2",
    "google-auth>=2.25.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.100.0",
    "notion-client",
    "aiohttp",
    "asyncpg",
    "redis>=5.0.0",
    "pyyaml",
    "alembic>=1.14.0",
]

[project.optional-dependencies]
browser = [
    "browser-use>=0.1.0",
    "langchain-google-genai>=2.0.0",
    "playwright>=1.40.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.25.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
all = [
    "cirkelline[browser,dev]",
]

[project.urls]
Homepage = "https://cirkelline.com"
Documentation = "https://docs.cirkelline.com"
Repository = "https://github.com/cirkelline/cirkelline-system"

[project.scripts]
cirkelline = "cirkelline.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["cirkelline*"]
exclude = ["tests*", "docs*", "ecosystems*", ".venv*", "CKC-COMPONENTS*"]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::pytest.PytestUnraisableExceptionWarning",
]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | ecosystems
    | __pycache__
    | build
    | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [
    ".git",
    ".venv",
    "ecosystems",
    "__pycache__",
    "build",
    "dist",
]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501", "B008"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = false
ignore_missing_imports = true
exclude = [
    "ecosystems/",
    ".venv/",
    "tests/",
]
